Temp fix for ipykernel >= 5, handle Future returned by do_one_iteration().
In ipykernel >= 5, do_one_iteration() has been changed into a coroutine and
returns a Future. This Future  needs to be processed appropriately, if not the
kernel will die with an error "execution aborted" (see #26) due to the coroutine
not having completed execution. To fix this, have ipython_kernel_iteration()
complete execution of the coroutine before continuing.

This is just a temporary fix, will need checks so the code works on
ipykernel < 5.